split that use JSX survey templates for rendering from code used by server and scripts
