+++ .vscode/settings.json
@@ -19,5 +19,11 @@
     "titleBar.inactiveForeground": "#e7e7e799",
     "commandCenter.border": "#e7e7e799"
   },
-  "peacock.color": "#272325"
+  "peacock.color": "#272325",
+  "search.exclude": {
+    "**/.next": true,
+    "**/.yarn": true,
+    "**/dist": true,
+    "**/pnpm-lock.yaml": true
+  }
 }
