+++ surveyform/package.json
@@ -241,8 +241,8 @@
   "packageManager": "pnpm@7.3.0",
   "pnpm": {
     "overrides": {
-      "@types/react": "$types/react",
-      "@types/react-dom": "$types/react-dom"
+      "@types/react>17": "^17",
+      "@types/react-dom>17": "^17"
     }
   }
 }
