+++ surveyform/src/admin/server/api/generateExports.ts
@@ -2,7 +2,7 @@ import { exec } from "child_process";
 import path from "path";
 import { serverConfig } from "~/config/server";
 import { captureMessage } from "@sentry/nextjs";
-import { SurveyDocument } from "~/surveys";
+import type { SurveyDocument } from "~/surveys";
 
 type SupportedFormat = "json" | "csv";
