+++ surveyform/src/core/components/survey/SurveyCredits.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useVulcanComponents } from "@vulcanjs/react-ui";
 import { useEntitiesQuery } from "~/core/hooks/useEntitiesQuery";
 import { Entity } from "~/modules/entities/typings";
-import { SurveyType } from "~/surveys";
+import type { SurveyType } from "~/surveys";
 
 const SurveyCredits = ({ survey }) => {
   const Components = useVulcanComponents();
