+++ surveyform/src/modules/surveys/outlineExport.ts
@@ -6,11 +6,9 @@
  */
 
 import {
-  Field,
   FieldTemplateId,
   ParsedQuestion,
   SurveyDocument,
-  SurveyQuestion,
   SurveySection,
 } from "~/surveys";
